add a dead stupid implementation of WriteFile
we allocate a single buffer that will be sent as content of the
object on flush. reads continue to read from the object directly
and not through this buffer, and read-modify-write is also not
implemented. it does enable us to do something useful with
command line utilities (cp, cat and the like)