Use full length env variable names.
Use 'AWS_ACCESS_KEY_ID' and 'AWS_SECRET_ACCESS_KEY' environment variable names
in README.md. These names match verbatim what the user sees in "Add
user", and correspond with other AWS utility configurations.
https://github.com/aws/aws-sdk-go#configuring-credentials

The underlying aws-sdk-go accepts the shorthand as a fallback, but if
the user has the full length variables set and goes to set the variables
listed in the docs thinking they are what Goofys uses, they will be in
for a suprise as the full length names take precedence. Changing the
docs to the full name- as this commit does- prevents that 'oops.'