optimize recursive listing of small dirs
when listing a subtree with many small directories, we were doing many
small ListObjects requests. ex:

a/
a/a/
a/a/a
a/a/b
a/b/
a/b/a
a/d

would require ListObjects(a), ListObjects(a/a/), and ListObjects(a/b/)
with delimiter set to /, even though just ListObjects(a) without
delimiter is sufficient.

We cannot apply this optimization if the subtree is > 1000 items, so we
can bound the amount of time readdir will take to complete.

With this change, listing my year/month/day/image_name.jpg collection
over internet improved > 2x:

$ time find /tmp/s3 >/dev/null

real    1m36.130s
user    0m0.072s
sys     0m0.052s

$ time find /tmp/s3 >/dev/null

real    0m39.645s
user    0m0.032s
sys     0m0.096s