workaround minio incompatibility
Another minio incompatibility: if object/ exists and no object/* exists,
ListObjects delimiter=/ prefix=object/ returns nothing, when we expect
it to return object/ itself. This leads us to try to rename object
instead of object/, and eventually panic because of other internal
confusion.

Not going to fix the rename problem but at least we won't crash.

Fixes #252