Fix ReadDir bug caused due to differences in how goofys and s3 orders objects (#58)
Problem: In s3 & azblob, prefixes are returned with '/' => the prefix "2019" is returned as "2019/". So the list api for these backends returns "2019/" after "2019-0001/" because ascii("/") > ascii("-"). This is problematic for goofys if "2019/" is returned in x+1'th batch and "2019-0001/" is returned in x'th; Goofys stores the results as they arrive in a sorted array and expects backends to return entries in a sorted order.
We cant just use ordering of s3/azblob because different cloud providers have different sorting strategies when it involes directories. In s3 "a/" > "a-b/". In adlv2 it is opposite.

Solution: To deal with this our solution with s3 & azblob is as follows. (nothing needed for others). For a single call of ListBlobs, we keep requesting multiple list batches until there is nothing left to list or the last listed entry has all characters > "/"